feat(step3): InputView 클래스와 상관없는 로직은 Utils로 이동
